# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Hong Tin Mang API",
    "description": "REST API cho sản phẩm, người dùng và đơn hàng. Cần cấu hình `MONGODB_URI` trên server.",
    "version": "1.0.0"
  },
  "servers": [{ "url": "/", "description": "Cùng origin với ứng dụng Next.js" }],
  "tags": [
    { "name": "products", "description": "Sản phẩm" },
    { "name": "users", "description": "Người dùng" },
    { "name": "orders", "description": "Đơn hàng" }
  ],
  "paths": {
    "/api/products": {
      "get": {
        "tags": ["products"],
        "summary": "Danh sách sản phẩm",
        "description": "Mặc định chỉ trả về sản phẩm đang bật (`active`). Thêm `all=1` để lấy tất cả (dùng cho admin).",
        "parameters": [
          {
            "name": "all",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "enum": ["1"] },
            "description": "Nếu bằng `1`, trả về cả sản phẩm không active."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "tags": ["products"],
        "summary": "Tạo sản phẩm",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProductCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đã tạo",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "tags": ["products"],
        "summary": "Chi tiết sản phẩm",
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } }
          },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      },
      "patch": {
        "tags": ["products"],
        "summary": "Cập nhật sản phẩm (partial)",
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProductPatch" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đã cập nhật",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } }
          },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "tags": ["products"],
        "summary": "Xóa sản phẩm",
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": {
            "description": "Đã xóa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "ok": { "type": "boolean", "example": true } },
                  "required": ["ok"]
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/users": {
      "get": {
        "tags": ["users"],
        "summary": "Danh sách người dùng",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/User" } }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "tags": ["users"],
        "summary": "Tạo người dùng",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đã tạo",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": ["users"],
        "summary": "Chi tiết người dùng",
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } }
          },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      },
      "patch": {
        "tags": ["users"],
        "summary": "Cập nhật người dùng (partial)",
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserPatch" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đã cập nhật",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } }
          },
          "404": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "tags": ["users"],
        "summary": "Xóa người dùng",
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": {
            "description": "Đã xóa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "ok": { "type": "boolean", "example": true } },
                  "required": ["ok"]
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/orders": {
      "get": {
        "tags": ["orders"],
        "summary": "Danh sách đơn hàng",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Order" } }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "tags": ["orders"],
        "summary": "Tạo đơn hàng",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OrderCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đã tạo",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Order" } } }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "tags": ["orders"],
        "summary": "Chi tiết đơn hàng",
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Order" } } }
          },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      },
      "patch": {
        "tags": ["orders"],
        "summary": "Cập nhật trạng thái đơn hàng",
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": ["pending", "confirmed", "delivered", "cancelled"]
                  }
                },
                "required": ["status"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đã cập nhật",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Order" } } }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "tags": ["orders"],
        "summary": "Xóa đơn hàng",
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": {
            "description": "Đã xóa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "ok": { "type": "boolean", "example": true } },
                  "required": ["ok"]
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "MongoDB ObjectId (chuỗi hex 24 ký tự)"
      }
    },
    "responses": {
      "Error": {
        "description": "Lỗi",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": { "error": { "type": "string" } },
              "required": ["error"]
            }
          }
        }
      }
    },
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "description": { "type": "string" },
          "price": { "type": "number" },
          "category": { "type": "string", "enum": ["breakfast", "cafe", "drink"] },
          "imageUrl": { "type": "string" },
          "active": { "type": "boolean" },
          "stock": { "type": "number" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["_id", "name", "slug", "price", "category", "active", "stock", "createdAt", "updatedAt"]
      },
      "ProductCreate": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "description": { "type": "string" },
          "price": { "type": "number", "minimum": 0 },
          "category": { "type": "string", "enum": ["breakfast", "cafe", "drink"] },
          "imageUrl": { "type": "string" },
          "active": { "type": "boolean" },
          "stock": { "type": "number" }
        },
        "required": ["name", "price", "category", "active", "stock"]
      },
      "ProductPatch": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "description": { "type": "string" },
          "price": { "type": "number", "minimum": 0 },
          "category": { "type": "string", "enum": ["breakfast", "cafe", "drink"] },
          "imageUrl": { "type": "string" },
          "active": { "type": "boolean" },
          "stock": { "type": "number" }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "name": { "type": "string" },
          "phone": { "type": "string" },
          "role": { "type": "string", "enum": ["user", "admin"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["_id", "email", "name", "role", "createdAt", "updatedAt"]
      },
      "UserCreate": {
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "name": { "type": "string" },
          "phone": { "type": "string" },
          "role": { "type": "string", "enum": ["user", "admin"] }
        },
        "required": ["email", "name", "role"]
      },
      "UserPatch": {
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "name": { "type": "string" },
          "phone": { "type": "string" },
          "role": { "type": "string", "enum": ["user", "admin"] }
        }
      },
      "OrderItemLine": {
        "type": "object",
        "properties": {
          "productId": { "type": "string" },
          "name": { "type": "string" },
          "price": { "type": "number" },
          "quantity": { "type": "number" }
        },
        "required": ["productId", "name", "price", "quantity"]
      },
      "Order": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "userId": { "type": "string" },
          "customerName": { "type": "string" },
          "customerPhone": { "type": "string" },
          "deliveryAddress": { "type": "string" },
          "note": { "type": "string" },
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/OrderItemLine" } },
          "total": { "type": "number" },
          "status": {
            "type": "string",
            "enum": ["pending", "confirmed", "delivered", "cancelled"]
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": [
          "_id",
          "customerName",
          "customerPhone",
          "deliveryAddress",
          "items",
          "total",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "OrderCreate": {
        "type": "object",
        "properties": {
          "userId": { "type": "string", "description": "Tùy chọn, ObjectId user" },
          "customerName": { "type": "string" },
          "customerPhone": { "type": "string" },
          "deliveryAddress": { "type": "string" },
          "note": { "type": "string" },
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/OrderItemLine" }, "minItems": 1 }
        },
        "required": ["customerName", "customerPhone", "deliveryAddress", "items"]
      }
    }
  }
}